ml-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
scipy==1.10.1
pymongo==4.6.1
python-dotenv==1.0.0
